move the helpers for figuring out the WT.exe path to a common location
